requirements.txt
PyMuPDF>=1.23.0
python-pptx>=0.6.21
python-docx>=0.8.11
anthropic>=0.18.0
google-generativeai==0.8.5
Pillow>=10.0.0
python-dotenv>=1.0.0
flask>=3.0.0
werkzeug>=3.0.0
